+++ src/containerapp/azext_containerapp/_models.py
@@ -587,3 +587,7 @@ CustomDomainConfiguration = {
     "certificatePassword": None,
     "certificateKeyVaultProperties": None
 }
+
+RuntimeJava = {
+    "enableMetrics": False
+}